Event model overhaul
There's now a clear separation between the structures by which the
game represents AI instructions, and those by which it represents
descriptions of events. The less obvious benefits of doing this
include the potential to (re-)unify all the AI types, in favor of
building a more flexible hierarchy (without rigid boundaries) from
individual, customized event responders. I've already laid some
groundwork for that goal with the Trigger type, and my immediate
concern is now to rebuild the main game loop around this concept.