Move AI EIDs into method closures
The reason I had these in the first place was that AI methods might
need access to entity data. But now that they're called in such a way
that they don't know what's calling them, that information has to go
directly into the closure that produces a responder for the map. That
makes the AI record's entity field totally redundant. I took it out.