Some entity attribute lookup functions
I'm in the middle of turning entity representation inside-out. Right
now, I have a giant record type that knows everything about an entity,
but needs to be updated all the time. When this change is complete,
there will instead be separate EID->Foo maps (as part of World) that
each track a certain attribute for every entity. This will have major
implications in the long run for how I sort and filter long lists of
entities (perhaps thousands at a time) without bogging down the game
by building and rebuilding useless intermediate data structures.

For the time being, I'm stopping here, because these changes will be
easier to make after certain other rewrites are in place. Namely, by
reifying event effects in the same way as textual descriptions of the
events in question, I'll reduce code duplication in the AI module and
ensure that any new kind of action I introduce can be described.